requirements.txt
# UI
streamlit>=1.40.0

# LLM
groq>=0.11.0

# Data
pandas>=2.2.0
openpyxl>=3.1.0
numpy>=1.26.0

# Visualization
plotly>=5.24.0

# Utils
python-dotenv>=1.0.0
fpdf2>=2.7.9
